qvm-template: use rpm2archive instead of rpm2cpio
The rpm2archive _does_ support writing to a pipe, not only to a file, so
it can be used here. rpm2archive deals better with various rpm formats
(especially with tar payload), but more importantly, does not rely on
LONGARCHIVESIZE tag that is zeroed by rpmcanon.

Fun fact: rpm2cpio worked on Fedora with rpmcanon only because Fedora
has a rpm2cpio patches to weaken payload size check.

QubesOS/qubes-issues#7080