Invalidate cache on 'connection-established' event
When admin.Events is called, it guarantees to deliver events only about
things happening after 'connection-established' event. Since there is no
guarantee about events delivery before that point, anything that was
cached before this point may not be updated/invalidated by relevant
event (which could be not delivered). Solve this by clearing the whole
cache on 'connection-established' event.

This applies both to initial connection, but also to re-connection to
qubesd.

Fixes QubesOS/qubes-issues#7362