backup: use ~/QubesIncoming as default restore tmpdir
SELinux policy is going to limit qfile-unpacker access to just this
directory. It makes sense, as the data indeed comes from a different VM.
Move the default restore location there, but use a directory name that
cannot conflict with an actual VM name (`#` is not valid in a VM name),
just to avoid unlikely corner cases.

Related to QubesOS/qubes-issues#4239