Invalidate cache on domain add/delete
When domain is removed, the VM names list cache was invalidated, but not
the actual objects. When same-named VM is later created, it might use
previously cached object that included properties cache. Fix this by
invalidating objects cache for specific object when domain is removed.
And for extra safety, when it's added too.

Fixes QubesOS/qubes-issues#7820