Update qvm-create.rst to fix markup issue `*NAME*=*VALUE*` (#7982)
The used technique is described in reStructuredText Markup Specification, chapter Character-Level Inline Markup: https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#toc-entry-46
Note, that this solution is not encouraged according to docs, should be used only when necessary.  I think it is exactly our case, because it seems there is no other way to get equal sign out of emphasis of the NAME and VALUE. Github preview of rst file is also correct now.
Alternative solutions would be to use `*NAME=VALUE*` or NAME=VALUE instead, but the result will not be completely right.