Fix: qvm-ls outputs to stderr for --help-* calls
The reason of the issue is that ArgumentParser has no ability to provide output file for exit() calls. In the ArgumentParser documentation [1] there is no information about output target file (stderr or stdout) for ArgumentParser.exit() functions, but from the source code of argparse [2] it is obvious that it always output message to stderr which is not desired in our case.
[1]: https://docs.python.org/3/library/argparse.html#exiting-methods
[2]: https://github.com/python/cpython/blob/3.11/Lib/argparse.py#L2612