Improve layout filtering of options for setxkbmap
Currently options like shift_caps_switch, lctrl_rctrl_switch and other first-last layout toggles is broken because they are not be filtered out.
The provided code maybe does it better. Note, that solution was not tested by me!
It is just an idea how it is possible to improve fc08f4e.

Docs: The list of options for keyboard layout switching is present in 'grp' chapter of `/usr/share/X11/xkb/rules`.
Note: I am not sure that "while pressed" shortcuts will work or if it work now.