tests: Account for argparse bug fix in Python 3.13
Test test_004_enable_opt_mixed is marked as xfail because of an argparse
bug, but the bug is fixed in Python 3.13 so the test unexpectedly
succeeds.  Remove the @unittest.expectedFailure, skip it on Python<3.13,
and add a new tests for older Python that checks for the previous buggy
behavior.

(cherry picked from commit 6461531d50e342ef4e6d030285f360a0c176a53b)