Ignore preload as a dependency
There doesn't seem to be a sane setup that would set a preloaded
disposable as a property of another qube or a system property. The core
already ignores preloaded disposables in certain cases, do the same on
the client to not have derived applications such as Qube Manager having
to deal with it.

Fixes: https://github.com/QubesOS/qubes-issues/issues/10227
For: https://github.com/QubesOS/qubes-issues/issues/1512