Cache GUIVM keyboard layout in `qvm-start-daemon`
On XFCE systems with multiple keyboard layouts and with XFCE Keyboard
Layout widget, X raises two `XKLAVIER_ALLOW_SECONDARY` events shortly
after each other (see 1st issue).

The current intended behaviour is to ignore the 2nd event by checking if
the GUIVMs current layout is similar to the new requested layout.
Unfortunately, after setting the GUIVM's keyboard layout, Qubes Core
goes to this loop to propagate the new layout to all child qubes:
https://github.com/QubesOS/qubes-core-admin/blob/main/qubes/ext/gui.py#L162-L170

And even if we move the condition after that loop just before it, it is
still too slow to report back the new layout via qubesdb for
`qvm-stat-daemon` use.

In order to avoid this race condition, it would be better to cache the
layout within `qvm-start-daemon` itself.

fixes: https://github.com/QubesOS/qubes-issues/issues/8441
fixes: https://github.com/QubesOS/qubes-issues/issues/6517